requirements.txt
absl-py==0.7.0
astor==0.7.1
atari-py==0.1.15
attrs==18.2.0
backcall==0.1.0
baselines==0.1.5
blessings==1.7
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
cloudpickle==0.6.1
ConfigArgParse==0.13.0
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.2
dask==1.0.0
decorator==4.3.0
descartes==1.1.0
dill==0.3.0
dm-control==0.0.295778102
dm-env==1.1
dm-sonnet==1.27
dm-tree==0.1.1
dominate==2.3.5
easydict==1.9
enum34==1.1.6
Fiona==1.8.4
future==0.17.1
futures==3.1.1
gast==0.2.2
geopandas==0.4.0
glfw==1.8.3
google-pasta==0.1.8
gpustat==0.6.0
graphviz==0.10.1
grpcio==1.18.0
gym==0.12.5
h5py==2.9.0
idna==2.8
imageio==2.4.1
imgaug==0.2.7
intel-openmp==2019.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
joblib==0.13.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lockfile==0.12.2
lxml==4.4.1
Markdown==3.0.1
matplotlib==3.0.2
mkl==2019.0
mkl-devel==2018.0.3
mkl-include==2019.0
mpi4py==3.0.2
mujoco-py==2.0.2.5
munch==2.3.2
networkx==2.2
numpy==1.15.4
nvidia-ml-py3==7.352.0
opencv-python==3.4.4.19
osmnx==0.8.2
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
progressbar2==3.39.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.6.3
ptyprocess==0.6.0
pycocotools==2.0.0
pycparser==2.19
pydot==1.4.1
pyglet==1.3.2
Pygments==2.3.1
pygobject==3.26.1
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.0
pyparsing==2.3.0
pypng==0.0.19
pyproj==1.9.6
pyssim==0.4
python-apt==1.6.3
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==18.1.0
rarfile==3.0
requests==2.21.0
Rtree==0.8.3
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
scikit-image==0.15.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
semantic-version==2.6.0
Shapely==1.6.4.post2
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
toolz==0.9.0
tqdm==4.28.1
traitlets==4.3.2
trimesh==2.35.49
typing==3.6.6
urllib3==1.24.1
virtualenv==16.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
zmq==0.0.0
